Reorganize new PAPI endpoint tests
