+++ tests/Clc.Polaris.Api.UnitTests/Methods/SortOptionsGetTests.cs
@@ -1,0 +1,19 @@
+namespace Clc.Polaris.Api.UnitTests.Methods
+{
+    [TestClass]
+    [UnitTest]
+    public sealed class SortOptionsGetTests : PapiClientUnitTestBase
+    {
+        [TestMethod]
+        public async Task SortOptionsGetAsync_SendsExpectedRequest()
+        {
+            var handler = new CapturingHttpMessageHandler(CreatePapiResponseJson());
+            var client = CreateClient(handler);
+
+            await client.SortOptionsGetAsync(branchId: 88, cancellationToken: TestContext.CancellationToken);
+
+            Assert.AreEqual(HttpMethod.Get, GetLastRequest(handler).Method);
+            Assert.AreEqual("/PAPIService/REST/public/v1/1033/100/88/sortoptions", GetLastRequestUri(handler).AbsolutePath);
+        }
+    }
+}
